Added delete incident confirm dialog. Added constraint to add more incidents. Remove scroll from search results

+++ src/IncidentDetails.js
@@ -102,8 +102,11 @@ class IncidentDetails extends React.Component {
 	}
 
 	onDeleteIncidentClick(e) {
-		e.preventDefault();
-		this.props.onDeleteIncidentClick(this.props.data);
+		var r = confirm("Tem a certeza que pretende eliminar?");
+		if (r === true) {
+			e.preventDefault();
+			this.props.onDeleteIncidentClick(this.props.data);
+		}
 	}
 
 	render() {
